move notification classes to the corresponding message-card components
