--- src/app/my-messages/abstract-community-notification.ts
@@ -1,6 +1,0 @@
-import { AbstractNotification } from './abstract-notification';
-
-export abstract class AbstractCommunityNotification extends AbstractNotification {
-
-  abstract getCommunityInfo(): string;
-}
