+++ src/app/my-messages/community-changed-message-card/community-details.enum.ts
