+++ src/app/my-messages/community-deleted-message-card/community-deleted-message-card.component.ts
@@ -2,7 +2,7 @@ import { Component, Input, OnInit } from '@angular/core';
 import { DeletableNotificationComponent } from '../deletable-notification/deletable-notification.component';
 import { MessagesService } from '../messages.service';
 import { MatDialog } from '@angular/material';
-import { CommunityDeletedNotification } from '../community-deleted-notification';
+import { CommunityDeletedNotification } from './community-deleted-notification';
 
 @Component({
   selector: 'app-community-deleted-message-card',
