+++ src/app/my-messages/community-invitation-message-card/community-invitation-notification.ts
@@ -1,7 +1,7 @@
-import { CommunityUserRegistrationResponse } from '../shared/community-user-registration-response';
-import { AbstractCommunityNotification } from './abstract-community-notification';
+import { CommunityUserRegistrationResponse } from '../../shared/community-user-registration-response';
+import { AbstractNotification } from '../abstract-notification';
 
-export class CommunityInvitationNotification extends AbstractCommunityNotification {
+export class CommunityInvitationNotification extends AbstractNotification {
   registration: CommunityUserRegistrationResponse;
   communityName: string;
