+++ src/app/my-messages/community-join-request-message-card/community-join-request-message-card.component.ts
@@ -2,7 +2,7 @@ import { Component, Input, OnInit } from '@angular/core';
 import { TodoNotificationComponent } from '../todo-notification/todo-notification.component';
 import { CommunityRegistrationService } from '../../shared/community-registration.service';
 import { MatDialog } from '@angular/material';
-import { JoinCommunityRequestNotification } from '../join-community-request-notification';
+import { JoinCommunityRequestNotification } from './join-community-request-notification';
 
 @Component({
   selector: 'app-community-join-request-message-card',
