+++ src/app/my-messages/community-join-request-message-card/join-community-request-notification.ts
@@ -1,7 +1,7 @@
-import { CommunityUserRegistrationResponse } from '../shared/community-user-registration-response';
-import { AbstractCommunityNotification } from './abstract-community-notification';
+import { CommunityUserRegistrationResponse } from '../../shared/community-user-registration-response';
+import { AbstractNotification } from '../abstract-notification';
 
-export class JoinCommunityRequestNotification extends AbstractCommunityNotification {
+export class JoinCommunityRequestNotification extends AbstractNotification {
   registration: CommunityUserRegistrationResponse;
   communityName: string;
